Move stateless functions from LocalStorageKeystore.ts to keystore.ts
